Use MUnit in OutboundStreamStateImplSuite
